Issue #62 - Allow http proxy values to be supplied through values.yaml
It is better to leave the JVM options to be configurable so that the end user can configure more options not just proxy.
Also providing env configuratble in values.yaml opens up opportunities for the users.
Updated the documentation to reflect the changes
Tested the image and helm chart in Kubernetes.